extensions for network client, basic refresh logic
